backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
sqlalchemy
psycopg2-binary
alembic
faiss-cpu
numpy 
python-multipart
PyPDF2

passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
pydantic[email]